Fixed bugs

+++ backend/src/test/java/com/kiwi/nodes/NodesTestFactory.java
@@ -9,7 +9,7 @@ import com.kiwi.features.nodes.data.UserNodeStatusKey;
 
 public class NodesTestFactory {
 
-    public static NodesPersistence persistenceNode(long id, int order, int price) {
+    public static NodesPersistence persistenceNode(int id, int order, int price) {
         NodesPersistence n = new NodesPersistence();
         n.setId(id);
         n.setOrder(order);
@@ -25,15 +25,15 @@ public class NodesTestFactory {
         return s;
     }
 
-    public static NodesDomain domainNode(long id, int order, NodeStatus status, int price) {
+    public static NodesDomain domainNode(int id, int order, NodeStatus status, int price) {
         return new NodesDomain(id, order, status, price);
     }
 
-    public static Object userIdDTO(Long id) {
+    public static Object userIdDTO(int id) {
         return new Object() { public int userId = id; };
     }
 
-    public static NodesDTO dtoNode(long id, int order, NodeStatus status, int price) {
+    public static NodesDTO dtoNode(int id, int order, NodeStatus status, int price) {
         return new NodesDTO(id, order, status.name(), price);
     }
